requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-filter==24.3
celery==5.4.0
redis==5.0.7
django-storages[boto3]==1.14.4
boto3==1.34.162
pgcrypto==2.10
psycopg2-binary==2.9.9
weasyprint==62.3
pandas==2.2.2
pytest==8.3.2
pytest-django==4.8.0
factory_boy==3.3.0
requests==2.32.3
